requirements.txt
Django==4.2.7
django-crispy-forms==2.1
crispy-bootstrap5==0.7
pandas==2.0.3
numpy==1.24.4
scikit-learn==1.3.2
Pillow==10.1.0
gunicorn